requirements.txt
addict==2.4.0
altair==4.1.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bcrypt==3.1.7
bleach==3.2.2
blinker==1.4
bokeh==2.2.3
boto3==1.16.61
botocore==1.19.61
Brlapi==0.7.0
Brotli==1.0.9
cachetools==4.2.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
confuse==1.4.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
htmlmin==0.1.12
httplib2==0.14.0
idna==2.8
ImageHash==4.2.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.35.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.3
missingno==0.4.2
mistune==0.8.4
monotonic==1.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.5
notebook==6.2.0
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
open3d==0.12.0
packaging==20.8
pandas==1.2.1
pandas-profiling==2.11.0
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
patsy==0.5.1
pexpect==4.6.0
phik==0.11.1
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2020.11.15
plotly==4.14.3
plotly-express==0.4.1
plyfile==0.7.3
prometheus-client==0.9.0
prompt-toolkit==3.0.13
protobuf==3.15.3
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==2.0.0
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydeck==0.5.0
Pygments==2.7.4
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynndescent==0.5.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==21.0.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
retrying==1.3.3
s3transfer==0.3.4
scikit-learn==0.24.1
scipy==1.6.0
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
smmap==3.0.5
statsmodels==0.12.2
streamlit==0.78.0
streamlit-pandas-profiling==0.1.1
systemd-python==234
tangled-up-in-unicode==0.0.6
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
umap-learn==0.5.0
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.18.2
virtualenv==20.4.0
virtualenv-clone==0.5.4
visions==0.6.0
wadllib==1.3.3
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xkit==0.0.0
